Implement readonly editor mode
