+++ src/features/editorMode/di.config.ts
@@ -1,8 +1,14 @@
 import { ContainerModule } from "inversify";
-import { TYPES, configureCommand } from "sprotty";
-import { ChangeEditorModeCommand, EditorModeController } from "./editorModeController";
+import { DeleteElementCommand, EditLabelMouseListener, MoveCommand, TYPES, configureCommand } from "sprotty";
+import { EditorModeController } from "./editorModeController";
 import { EditorModeSwitchUi } from "./modeSwitchUi";
 import { EDITOR_TYPES } from "../../utils";
+import {
+    EditorModeAwareDeleteElementCommand,
+    EditorModeAwareEditLabelMouseListener,
+    EditorModeAwareMoveCommand,
+} from "./sprottyHooks";
+import { ChangeEditorModeCommand } from "./command";
 
 export const editorModeModule = new ContainerModule((bind, unbind, isBound, rebind) => {
     const context = { bind, unbind, isBound, rebind };
@@ -13,4 +19,10 @@ export const editorModeModule = new ContainerModule((bind, unbind, isBound, rebi
     bind(EDITOR_TYPES.DefaultUIElement).toService(EditorModeSwitchUi);
 
     configureCommand(context, ChangeEditorModeCommand);
+
+    // Sprotty hooks that hook into the edit label, move and edit module
+    // to intercept model modifications to prevent them when the editor is in a read-only mode.
+    rebind(EditLabelMouseListener).to(EditorModeAwareEditLabelMouseListener).inSingletonScope();
+    rebind(MoveCommand).to(EditorModeAwareMoveCommand).inSingletonScope();
+    rebind(DeleteElementCommand).to(EditorModeAwareDeleteElementCommand).inSingletonScope();
 });
